backend/question table/adding question/requirements.txt
google-cloud-firestore==2.3.2
google-cloud-language==2.4.0

